requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
python-json-logger
pydantic-settings
